feat(web): schedule studio shell behind ?v=2
